refactor: refactor package names (#4)
