+++ BE/settings.gradle.kts
@@ -1,1 +1,1 @@
-rootProject.name = "be"
+rootProject.name = "antalbum"
