+++ BE/src/main/java/ssafy/antalbum/AntalbumApplication.java
@@ -1,13 +1,13 @@
-package ssafy.antalbum.be;
+package ssafy.antalbum;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class BeApplication {
+public class AntalbumApplication {
 
 	public static void main(String[] args) {
-		SpringApplication.run(BeApplication.class, args);
+		SpringApplication.run(AntalbumApplication.class, args);
 	}
 
 }
