+++ BE/src/test/java/ssafy/antalbum/AntalbumApplicationTests.java
@@ -1,10 +1,10 @@
-package ssafy.antalbum.be;
+package ssafy.antalbum;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-class BeApplicationTests {
+class AntalbumApplicationTests {
 
 	@Test
 	void contextLoads() {
